# contracts/openapi-before.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/atracciones/search": {
      "post": {
        "operationId": "AtraccionesController_search",
        "summary": "Buscar atracciones",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchAtraccionesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchAtraccionesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO de búsqueda inválido."
          }
        },
        "tags": [
          "Atracciones"
        ]
      }
    },
    "/api/v1/atracciones/details": {
      "post": {
        "operationId": "AtraccionesController_getDetailsBatch",
        "summary": "Obtener detalles batch de atracciones",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetailsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchAtraccionesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO de detalles inválido."
          }
        },
        "tags": [
          "Atracciones"
        ]
      }
    },
    "/api/v1/atracciones": {
      "post": {
        "operationId": "AtraccionesController_create",
        "summary": "Crear atracción",
        "description": "Requiere Bearer JWT y rol ADMIN.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAtraccionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Atracción creada; incluye cabecera Location.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtraccionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO inválido."
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "Se requiere rol ADMIN."
          }
        },
        "tags": [
          "Atracciones"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "AtraccionesController_findAll",
        "summary": "Listar atracciones paginadas",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página actual",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página",
            "schema": {
              "minimum": 1,
              "default": 100,
              "type": "number"
            }
          },
          {
            "name": "product_type",
            "required": false,
            "in": "query",
            "description": "Tipo de producto",
            "schema": {
              "enum": [
                "SINGLE_TICKET",
                "GUIDED_TOUR",
                "PACKAGE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtraccionesListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Parámetros de paginación o modalidad inválidos."
          }
        },
        "tags": [
          "Atracciones"
        ]
      }
    },
...
    "/api/v1/atracciones/{id}/paquetes": {
      "get": {
        "operationId": "AtraccionesController_getPackages",
        "summary": "Listar paquetes, modalidades, precios y políticas de una atracción",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID de la atracción.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaqueteExperienciaResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "El id no es un UUID válido."
          },
          "404": {
            "description": "Atracción no encontrada."
          }
        },
        "tags": [
          "Atracciones"
        ]
      }
    },
    "/api/v1/atracciones/reservations": {
      "get": {
        "operationId": "AtraccionesController_getReservations",
        "summary": "Listar mis reservas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente o inválido."
          }
        },
        "tags": [
          "Atracciones"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/atracciones/reservations/{reservationId}": {
      "get": {
        "operationId": "AtraccionesController_getReservationById",
        "summary": "Consultar detalle de una reserva propia",
        "parameters": [
          {
            "name": "reservationId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "El id no es un UUID válido."
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "La reserva pertenece a otro usuario."
          },
          "404": {
            "description": "Reserva no encontrada."
          }
        },
        "tags": [
          "Atracciones"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/atracciones/{id}": {
      "get": {
        "operationId": "AtraccionesController_getAttraction",
        "summary": "Consultar detalle de una atracción",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtraccionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "El id no es un UUID válido."
          },
          "404": {
            "description": "Atracción no encontrada."
          }
        },
        "tags": [
          "Atracciones"
        ]
      },
      "put": {
        "operationId": "AtraccionesController_replace",
        "summary": "Reemplazar una atracción",
        "description": "Requiere Bearer JWT y rol ADMIN.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAtraccionDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Reemplazo completado sin cuerpo."
          },
          "400": {
            "description": "UUID o DTO inválido."
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "Se requiere rol ADMIN."
          },
          "404": {
            "description": "Atracción no encontrada."
          }
        },
        "tags": [
          "Atracciones"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "AtraccionesController_update",
        "summary": "Actualizar parcialmente una atracción",
        "description": "Requiere Bearer JWT y rol ADMIN.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAtraccionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtraccionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "UUID o DTO inválido."
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "Se requiere rol ADMIN."
          },
          "404": {
            "description": "Atracción no encontrada."
          }
        },
        "tags": [
          "Atracciones"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "AtraccionesController_remove",
        "summary": "Desactivar una atracción (soft delete)",
        "description": "Requiere Bearer JWT y rol ADMIN.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Atracción desactivada sin cuerpo."
          },
          "400": {
            "description": "El id no es un UUID válido."
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "Se requiere rol ADMIN."
          },
          "404": {
            "description": "Atracción no encontrada."
          }
        },
        "tags": [
          "Atracciones"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/atracciones/{id}/availability": {
      "get": {
        "operationId": "AtraccionesController_getAvailability",
        "summary": "Consultar disponibilidad para una fecha y turno",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Fecha en formato YYYY-MM-DD.",
            "schema": {
              "example": "2026-10-10",
              "type": "string"
            }
          },
          {
            "name": "product_type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "SINGLE_TICKET",
                "GUIDED_TOUR",
                "PACKAGE"
              ],
              "type": "string"
            }
          },
          {
            "name": "time",
            "required": false,
            "in": "query",
            "description": "Si se envía, available_spots corresponde a este turno.",
            "schema": {
              "example": "10:00",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve date, available_spots, times y los filtros enviados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Fecha faltante/incorrecta o modalidad inválida."
          },
          "404": {
            "description": "Atracción no encontrada."
          }
        },
        "tags": [
          "Atracciones"
        ]
      }
    },
    "/api/v1/atracciones/{id}/reservations": {
      "post": {
        "operationId": "AtraccionesController_reserve",
        "summary": "Checkout y creación de reserva",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID de la atracción.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "UUID v4 estable para reintentos idénticos.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva confirmada y pago mock procesado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO/header invalido, paquete de otra atraccion o product_type diferente del paquete seleccionado."
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "404": {
            "description": "Atraccion o paquete_id no encontrado."
          },
          "409": {
            "description": "409 INSUFFICIENT_AVAILABILITY, PAYMENT_FAILED o IDEMPOTENCY_KEY_REUSED si la key tiene otro payload/propietario."
          }
        },
        "tags": [
          "Atracciones"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/atracciones/reservations/{reservationId}/cancel": {
      "post": {
        "operationId": "AtraccionesController_cancelReservation",
        "summary": "Cancelar una reserva propia",
        "parameters": [
          {
            "name": "reservationId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "UUID v4 para reintentos de cancelación.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelReservationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO inválido o header ausente/inválido."
          },
          "401": {
            "description": "JWT ausente o inválido."
          },
          "403": {
            "description": "La reserva pertenece a otro usuario."
          },
          "404": {
            "description": "Reserva no encontrada."
          },
          "409": {
            "description": "La key corresponde a otra cancelación o los cupos son inconsistentes."
          }
        },
        "tags": [
          "Atracciones"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Booking Prototipo API",
    "description": "API activa del módulo de atracciones. Errores: 400 (validación), 401 (JWT), 403 (rol o propiedad), 404 (recurso) y 409 (disponibilidad, pago o idempotencia). Las rutas administrativas requieren Bearer JWT y rol ADMIN.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "SearchAtraccionesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AtraccionResponseDto"
            }
          },
          "metadata": {
            "$ref": "#/components/schemas/SearchMetadataDto"
          },
          "request_id": {
            "type": "string",
            "example": "req-1720000000000"
          }
        },
        "required": [
          "data",
          "metadata",
          "request_id"
        ]
      },
      "DetailsRequestDto": {
        "type": "object",
        "properties": {
          "attractions": {
            "description": "Array de IDs de atracciones",
            "example": [
              "PRahAzWtTraa"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "languages": {
            "description": "Idiomas solicitados",
            "example": [
              "en-gb"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "product_type": {
            "type": "string",
            "description": "Filtra por tipo de producto",
            "enum": [
              "SINGLE_TICKET",
              "GUIDED_TOUR",
              "PACKAGE"
            ],
            "example": "GUIDED_TOUR"
          }
        },
        "required": [
          "attractions"
        ]
      },
      "BatchAtraccionesResponseDto": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "example": "batch-1720000000000"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AtraccionResponseDto"
            }
          }
        },
        "required": [
          "request_id",
          "data"
        ]
      },
...
    }
  }
}